[feature] it's not renders

+++ include/siren/bresenham_manager.hpp
@@ -9,20 +9,19 @@
 class BresenhamManager
 {
  private:
-  int x0;
-  int x1;
-  int y0;
-  int y1;
-  tga_image_t *image;
+  float x0;
+  float x1;
+  float y0;
+  float y1;
+  tga_image_t* image;
   tga_color_t color;
 
  public:
-  BresenhamManager(tga_image_t *image);
-  BresenhamManager(int x0, int y0, int x1, int y1, tga_image_t *image, tga_color_t color);
-  void setEndpoints(int x0, int y0, int x1, int y1);
+  BresenhamManager(tga_image_t* image);
+  BresenhamManager(float x0, float y0, float x1, float y1, tga_image_t* image, tga_color_t color);
+  void setEndpoints(float x0, float y0, float x1, float y1);
   void setColor(tga_color_t color);
-  void setImage(tga_image_t *image);
+  void setImage(tga_image_t* image);
   void drawLine();
 };
-
 #endif  // BRESENHAM_MANAGER_HPP_
